WIP: Extract async request handling code
