+++ hishel/_async/_transports.py
@@ -5,14 +5,13 @@ import typing as tp
 
 import httpcore
 import httpx
-from httpx import AsyncByteStream, Request, Response
-from httpx._exceptions import ConnectError
+from httpx import AsyncByteStream
 
-from hishel._utils import extract_header_values_decoded, normalized_url
+from hishel._utils import normalized_url
 
-from .._controller import Controller, allowed_stale
-from .._headers import parse_cache_control
-from .._serializers import JSONSerializer, Metadata
+from .._controller import Controller
+from .._serializers import JSONSerializer
+from ._request_handler import AsyncCacheRequestHandler
 from ._storages import AsyncBaseStorage, AsyncFileStorage
 
 if tp.TYPE_CHECKING:  # pragma: no cover
@@ -21,12 +20,36 @@ if tp.TYPE_CHECKING:  # pragma: no cover
 __all__ = ("AsyncCacheTransport",)
 
 
-async def fake_stream(content: bytes) -> tp.AsyncIterable[bytes]:
-    yield content
+HTTPX_EXC_TO_HTTPCORE_EXC = {
+    httpx.TimeoutException: httpcore.TimeoutException,
+    httpx.ConnectTimeout: httpcore.ConnectTimeout,
+    httpx.ReadTimeout: httpcore.ReadTimeout,
+    httpx.WriteTimeout: httpcore.WriteTimeout,
+    httpx.PoolTimeout: httpcore.PoolTimeout,
+    httpx.NetworkError: httpcore.NetworkError,
+    httpx.ConnectError: httpcore.ConnectError,
+    httpx.ReadError: httpcore.ReadError,
+    httpx.WriteError: httpcore.WriteError,
+    httpx.ProxyError: httpcore.ProxyError,
+    httpx.UnsupportedProtocol: httpcore.UnsupportedProtocol,
+    httpx.ProtocolError: httpcore.ProtocolError,
+    httpx.LocalProtocolError: httpcore.LocalProtocolError,
+    httpx.RemoteProtocolError: httpcore.RemoteProtocolError,
+}
+# Same thing:
+# TODO: Decide.
+# HTTPX_EXC_TO_HTTPCORE_EXC = {
+#     httpx_exc: httpcore_exc
+#     for httpcore_exc, httpx_exc
+#     in httpx._transports.HTTPCORE_EXC_MAP.items()
+# }
 
 
-def generate_504() -> Response:
-    return Response(status_code=504)
+def make_httpcore_exc_from_httpx_exc(httpx_exc: httpx.TransportError) -> Exception:
+    httpcore_exc_type = HTTPX_EXC_TO_HTTPCORE_EXC.get(type(httpx_exc))
+    if httpcore_exc_type is not None:
+        return httpcore_exc_type(httpx_exc)
+    return httpx_exc
 
 
 class AsyncCacheStream(AsyncByteStream):
@@ -62,6 +85,7 @@ class AsyncCacheTransport(httpx.AsyncBaseTransport):
     ) -> None:
         self._transport = transport
 
+        # TODO: Maybe pass storage and controller to the request handler.
         self._storage = storage if storage is not None else AsyncFileStorage(serializer=JSONSerializer())
 
         if not isinstance(self._storage, AsyncBaseStorage):
@@ -69,7 +93,36 @@ class AsyncCacheTransport(httpx.AsyncBaseTransport):
 
         self._controller = controller if controller is not None else Controller()
 
-    async def handle_async_request(self, request: Request) -> Response:
+        self._request_handler = AsyncCacheRequestHandler(
+            controller=self._controller,
+            storage=self._storage,
+            base_request_handler=self._handle_async_httpcore_request,
+        )
+
+    async def _handle_async_httpcore_request(self, request: httpcore.Request) -> httpcore.Response:
+        assert isinstance(request.stream, tp.AsyncIterable)
+        httpx_request = httpx.Request(
+            method=request.method,
+            url=normalized_url(request.url),
+            headers=request.headers,
+            stream=AsyncCacheStream(request.stream),
+            extensions=request.extensions,
+        )
+        try:
+            httpx_response = await self._transport.handle_async_request(httpx_request)
+        except httpx.TransportError as exc:
+            raise make_httpcore_exc_from_httpx_exc(exc) from exc
+
+        assert isinstance(httpx_response.stream, tp.AsyncIterable)
+        httpcore_response = httpcore.Response(
+            status=httpx_response.status_code,
+            headers=httpx_response.headers.raw,
+            content=AsyncCacheStream(httpx_response.stream),
+            extensions=httpx_response.extensions,
+        )
+        return httpcore_response
+
+    async def handle_async_request(self, request: httpx.Request) -> httpx.Response:
         """
         Handles HTTP requests while also implementing HTTP caching.
 
@@ -78,25 +131,6 @@ class AsyncCacheTransport(httpx.AsyncBaseTransport):
         :return: An HTTP response
         :rtype: httpx.Response
         """
-
-        if request.extensions.get("cache_disabled", False):
-            request.headers.update(
-                [
-                    ("Cache-Control", "no-store"),
-                    ("Cache-Control", "no-cache"),
-                    *[("cache-control", value) for value in request.headers.get_list("cache-control")],
-                ]
-            )
-
-        if request.method not in ["GET", "HEAD"]:
-            # If the HTTP method is, for example, POST,
-            # we must also use the request data to generate the hash.
-            body_for_key = await request.aread()
-            request.stream = AsyncCacheStream(fake_stream(body_for_key))
-        else:
-            body_for_key = b""
-
-        # Construct the HTTPCore request because Controllers and Storages work with HTTPCore requests.
         httpcore_request = httpcore.Request(
             method=request.method,
             url=httpcore.URL(
@@ -109,155 +143,12 @@ class AsyncCacheTransport(httpx.AsyncBaseTransport):
             content=request.stream,
             extensions=request.extensions,
         )
-
-        key = self._controller._key_generator(httpcore_request, body_for_key)
-        stored_data = await self._storage.retrieve(key)
-
-        request_cache_control = parse_cache_control(
-            extract_header_values_decoded(request.headers.raw, b"Cache-Control")
-        )
-
-        if request_cache_control.only_if_cached and not stored_data:
-            return generate_504()
-
-        if stored_data:
-            # Try using the stored response if it was discovered.
-
-            stored_response, stored_request, metadata = stored_data
-
-            # Immediately read the stored response to avoid issues when trying to access the response body.
-            stored_response.read()
-
-            res = self._controller.construct_response_from_cache(
-                request=httpcore_request,
-                response=stored_response,
-                original_request=stored_request,
-            )
-
-            if isinstance(res, httpcore.Response):
-                # Simply use the response if the controller determines it is ready for use.
-                return await self._create_hishel_response(
-                    key=key,
-                    response=res,
-                    request=httpcore_request,
-                    cached=True,
-                    revalidated=False,
-                    metadata=metadata,
-                )
-
-            if request_cache_control.only_if_cached:
-                return generate_504()
-
-            if isinstance(res, httpcore.Request):
-                # Controller has determined that the response needs to be re-validated.
-                assert isinstance(res.stream, tp.AsyncIterable)
-                revalidation_request = Request(
-                    method=res.method,
-                    url=normalized_url(res.url),
-                    headers=res.headers,
-                    stream=AsyncCacheStream(res.stream),
-                    extensions=res.extensions,
-                )
-                try:
-                    revalidation_response = await self._transport.handle_async_request(revalidation_request)
-                except ConnectError:
-                    # If there is a connection error, we can use the stale response if allowed.
-                    if self._controller._allow_stale and allowed_stale(response=stored_response):
-                        return await self._create_hishel_response(
-                            key=key,
-                            response=stored_response,
-                            request=httpcore_request,
-                            cached=True,
-                            revalidated=False,
-                            metadata=metadata,
-                        )
-                    raise  # pragma: no cover
-                assert isinstance(revalidation_response.stream, tp.AsyncIterable)
-                httpcore_revalidation_response = httpcore.Response(
-                    status=revalidation_response.status_code,
-                    headers=revalidation_response.headers.raw,
-                    content=AsyncCacheStream(revalidation_response.stream),
-                    extensions=revalidation_response.extensions,
-                )
-
-                # Merge headers with the stale response.
-                final_httpcore_response = self._controller.handle_validation_response(
-                    old_response=stored_response,
-                    new_response=httpcore_revalidation_response,
-                )
-
-                await final_httpcore_response.aread()
-                await revalidation_response.aclose()
-
-                assert isinstance(final_httpcore_response.stream, tp.AsyncIterable)
-
-                # RFC 9111: 4.3.3. Handling a Validation Response
-                # A 304 (Not Modified) response status code indicates that the stored response can be updated and
-                # reused. A full response (i.e., one containing content) indicates that none of the stored responses
-                # nominated in the conditional request are suitable. Instead, the cache MUST use the full response to
-                # satisfy the request. The cache MAY store such a full response, subject to its constraints.
-                if revalidation_response.status_code != 304 and self._controller.is_cachable(
-                    request=httpcore_request, response=final_httpcore_response
-                ):
-                    await self._storage.store(key, response=final_httpcore_response, request=httpcore_request)
-
-                return await self._create_hishel_response(
-                    key=key,
-                    response=final_httpcore_response,
-                    request=httpcore_request,
-                    cached=revalidation_response.status_code == 304,
-                    revalidated=True,
-                    metadata=metadata,
-                )
-
-        regular_response = await self._transport.handle_async_request(request)
-        assert isinstance(regular_response.stream, tp.AsyncIterable)
-        httpcore_regular_response = httpcore.Response(
-            status=regular_response.status_code,
-            headers=regular_response.headers.raw,
-            content=AsyncCacheStream(regular_response.stream),
-            extensions=regular_response.extensions,
-        )
-        await httpcore_regular_response.aread()
-        await httpcore_regular_response.aclose()
-
-        if self._controller.is_cachable(request=httpcore_request, response=httpcore_regular_response):
-            await self._storage.store(
-                key,
-                response=httpcore_regular_response,
-                request=httpcore_request,
-            )
-
-        return await self._create_hishel_response(
-            key=key,
-            response=httpcore_regular_response,
-            request=httpcore_request,
-            cached=False,
-            revalidated=False,
-        )
-
-    async def _create_hishel_response(
-        self,
-        key: str,
-        response: httpcore.Response,
-        request: httpcore.Request,
-        cached: bool,
-        revalidated: bool,
-        metadata: Metadata | None = None,
-    ) -> Response:
-        if cached:
-            assert metadata
-            metadata["number_of_uses"] += 1
-            await self._storage.update_metadata(key=key, request=request, response=response, metadata=metadata)
-            response.extensions["from_cache"] = True  # type: ignore[index]
-            response.extensions["cache_metadata"] = metadata  # type: ignore[index]
-        else:
-            response.extensions["from_cache"] = False  # type: ignore[index]
-        response.extensions["revalidated"] = revalidated  # type: ignore[index]
-        return Response(
+        response = await self._request_handler.handle_async_request(httpcore_request)
+        assert isinstance(response.stream, tp.AsyncIterable)
+        return httpx.Response(
             status_code=response.status,
             headers=response.headers,
-            stream=AsyncCacheStream(fake_stream(response.content)),
+            stream=AsyncCacheStream(response.stream),
             extensions=response.extensions,
         )
